requirements.txt
Django>=5.0
python-dotenv
Pillow
PyMuPDF
pdfminer.six
python-docx
groq
PyPDF2
pytesseract